reworking some things to be generally better - fingers crossed this doesn't break anything - moving IIRFilter to its own .py file and adding some properties. Also removing matplotlib as a dependency since it can cause some issues and really isn't needed.
